Move out tensor storage into a separate .hpp/.cpp
