+++ ttnn/cpp/ttnn/tensor/CMakeLists.txt
@@ -3,6 +3,7 @@ set(TENSOR_SRCS
     ${CMAKE_CURRENT_SOURCE_DIR}/tensor.cpp
     ${CMAKE_CURRENT_SOURCE_DIR}/tensor_ops.cpp
     ${CMAKE_CURRENT_SOURCE_DIR}/types.cpp
+    ${CMAKE_CURRENT_SOURCE_DIR}/storage.cpp
     ${CMAKE_CURRENT_SOURCE_DIR}/tensor_utils.cpp
     ${CMAKE_CURRENT_SOURCE_DIR}/serialization.cpp
     ${CMAKE_CURRENT_SOURCE_DIR}/shape/shape_base.cpp
