+++ ttnn/cpp/ttnn/tensor/storage.cpp
@@ -1,0 +1,19 @@
+// SPDX-FileCopyrightText: © 2025 Tenstorrent Inc.
+//
+// SPDX-License-Identifier: Apache-2.0
+
+#include "ttnn/tensor/storage.hpp"
+
+namespace tt::tt_metal {
+
+std::vector<DeviceBuffer> MultiDeviceStorage::get_buffers() const {
+    std::lock_guard<std::mutex> lock(buffer_mtx);
+    std::vector<DeviceBuffer> buf_vec;
+    buf_vec.reserve(buffers.size());
+    for (const auto& pair : buffers) {
+        buf_vec.push_back(pair.second);
+    }
+    return buf_vec;
+}
+
+}  // namespace tt::tt_metal
